+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
 - Append new items to make git merging easier.
+- feat(autocompleter): fetch preloaded value to ensure value is always rendered
 - feat(select-cell): implement onChange API
 - feat(input-cell): implement onChange API